+++ CHANGE_LOG.md
@@ -1,3 +1,27 @@
+# [1.244.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.243.0...v1.244.0) (2023-09-19)
+
+
+### Bug Fixes
+
+* **homepage:** merge main ([97f2690](https://github.com/oaknational/Oak-Web-Application/commit/97f2690790ff5cafce56709977ee7997fc6e55eb))
+* **homepage:** remove comments increase test coverage ([b139f7f](https://github.com/oaknational/Oak-Web-Application/commit/b139f7f884344a03cdf6d3aa82fdfbb2737a75b6))
+* **homepage:** remove dupe code ([919b830](https://github.com/oaknational/Oak-Web-Application/commit/919b830a2b924fb32ef6f1c43f33d6e064011509))
+* **homepage:** style and accessibility fixes ([54b90c1](https://github.com/oaknational/Oak-Web-Application/commit/54b90c1dacd0b6dec5838ed796d79c4b3f4ceb4e))
+* **homepage:** test coverage increase ([a877cfa](https://github.com/oaknational/Oak-Web-Application/commit/a877cfa5bb2e6f6e3de8cb66db688e476f96f290))
+* **homepage:** test fix ([b1e22a9](https://github.com/oaknational/Oak-Web-Application/commit/b1e22a93f72b35251c8156bc04a1ac34cc87e40a))
+* **homepage:** test fix and fix icon ([bce0f1d](https://github.com/oaknational/Oak-Web-Application/commit/bce0f1d2ba23e11926c25ac3b437f0390e556afd))
+
+
+### Features
+
+* **home-page-tabs:** add homepage tab to storybook ([f95639a](https://github.com/oaknational/Oak-Web-Application/commit/f95639a31837393f7cece835396ef2ac594fcbcf))
+* **home-page-tabs:** add new sanity assets ([8ed0f8a](https://github.com/oaknational/Oak-Web-Application/commit/8ed0f8a5600e833cc2b07573662b75a5f8fc69c9))
+* **home-page-tabs:** add new tab to teachers home and move old beta home to storybook ([b8f6b77](https://github.com/oaknational/Oak-Web-Application/commit/b8f6b7738d6c094027a7685f87a4d4617e18e87a))
+* **home-page-tabs:** create homepage tab and mount on storybook ([d58164e](https://github.com/oaknational/Oak-Web-Application/commit/d58164ed13e09bce4b0b2959ee3720e7cab6d6ee))
+* **home-page-tabs:** merge main ([e96d79c](https://github.com/oaknational/Oak-Web-Application/commit/e96d79ce33f712f9bbc5d6099f57989ea7317687))
+* **home-page-tabs:** remove old beta home from homepage ([02748db](https://github.com/oaknational/Oak-Web-Application/commit/02748dba1f8975078b4a7a9ec44d24bdd89eacb2))
+* **home-page-tabs:** style teachers tab ([deed667](https://github.com/oaknational/Oak-Web-Application/commit/deed667c6cab2a8e5367d803556b91cb9a14b067))
+
 # [1.243.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.242.0...v1.243.0) (2023-09-18)